Circular Mirror simuation
This program models the trajectory of a particle rebounding off the interior of a circular room with no energy losses
The main turtle (drawing the particle) is assigned a random coordinate within the mirror and assigned a random heading. It detects the presence of the mirror by tracking its own coordinate being within the equation of the circle making the mirror. When the mirror is detected, the heading is changed, obeying the law of reflection, calculated using calculus and finding the angle to turn anticlockwise. This program runs forever using a while True loop