Elliptical Snooker (endless)
Simulation of an elliptical snooker table where one focus is the pot and the other is the starting point of the cue ball. This simulation models the random straight line trajectory of the cue ball, proving that it indeed always ends up in the other focus. The trials endlessly repeat in a while True loop.